Can find more than a single link for file
Originally due to a bug where search stops at the first match it finds
and the fact that the entire file was being treated as one string only a
max of one link would be found for each file. Currently the hotfix to to
consider each line individually but that creates an issue of it there is
more than one link for line.